Moved tests to a seperate file
